DOCPN-CIM/requirements.txt
torch 
torchvision 
torchaudio 
torchtext
diffusers
transformers 
scipy 
ftfy
ultralytics
ipywidgets==7.7.2s
accelerate
efficientnet_pytorch
ipython-autotime
yolov5
sewar
image_similarity_measures
sentence_transformers
ftfy
regex
sentencepiece
keras
Markdown
MarkupSafe
matplotlib
pathlib
pycocotools
requests
seaborn
tqdm
pillow 
spacy 
pandas 
json 
opencv-python
Cython 
nltk 
huggingface_hub 
numpy 
scikit-image 
scikit-learn 
statistics 
tensorflow 
efficientnet_pytorch
beautifulsoup4
tabulate
tensorflow_hub
termcolor
iopath
cloudpickle
flair
docutils
ipython
PyYAML
recommonmark
wordcloud